(feature) Create EffectsViewSet
- Use faker to assign random file names to be used when photo preview
files are being created on disk. Random file names are preferred so as
to trigger a refresh on the client side whenever new thumbnails are
created.
- Create logic to handle cases where the 'path' argument is a string
(a file's url) instead of an instance of InMemoryUploadedFile. When a
file url string is used, a file object is created from that and then
passed on to the serializer through the data dictionary
- When POST is called on this route, the number of thumbnails created
will correspond to the number of effects/filters this app currently
has to offer